[Patchset 12] Remove warning suppressions and fix many casting bugs
This change removes most warning suppressions, in particular
from production code. In the past we inherit a bunch of -Wno-xxx
suppressions required to build gtest and libprotobuf headers.
Doing so, however, caused the suppressions to propagate back to
the translation units that were including any protobuf header
or any auto-generated .pb.h stub.
This change moves the gtest and probobuf header to be a
system include (-isystem vs -I). Doing so implicitly blacklists
any compiler warning on the headers.
This CL then re-enables warnings and deals with the fall out of
fixes that came out of this.

Bug: 77316877
Test: pefetto_unittests / perfetto_integrationtests

+++ src/tracing/core/service_impl.cc
@@ -981,7 +981,7 @@ TraceBuffer* ServiceImpl::GetBufferByID(BufferID buffer_id) {
 void ServiceImpl::UpdateMemoryGuardrail() {
 #if !PERFETTO_BUILDFLAG(PERFETTO_CHROMIUM_BUILD) && \
     !PERFETTO_BUILDFLAG(PERFETTO_OS_MACOSX)
-  size_t total_buffer_bytes = 0;
+  uint64_t total_buffer_bytes = 0;
 
   // Sum up all the shared memory buffers.
   for (const auto& id_to_producer : producers_) {
@@ -996,8 +996,7 @@ void ServiceImpl::UpdateMemoryGuardrail() {
 
   // Set the guard rail to 32MB + the sum of all the buffers over a 30 second
   // interval.
-  uint32_t guardrail =
-      32 * 1024 * 1024 + static_cast<uint32_t>(total_buffer_bytes);
+  uint64_t guardrail = 32 * 1024 * 1024 + total_buffer_bytes;
   base::Watchdog::GetInstance()->SetMemoryLimit(guardrail, 30 * 1000);
 #endif
 }
